// src/models/user.ts
export interface User {
  id: number;
  familyName: string;
  givenName: string;
  familyNameKana: string;
  givenNameKana: string;
  postalCode: string;
  prefecture: string;
  address1: string;
  address2: string;
  address3: string;
  phoneNumber: string;
  homePhoneNumber: string;
  email: string;
  gender: number;
  birthday: string;
  occupation: string;
  firstVisitDate: string;
  dmForwardingFlg: string;
  memo: string;
  reasonForComing: string;
  familyUserId: number;
  nextVisitDate: string;
  previousVisitDate: string;
  lastVistDates: number;
  avatar: string;
  name: string;
  jobtitle: string;
}